move resource callbacks to the resources submodule of mapper
